Moved classes to namespace
