+++ Where_Is_My_Wife/Assets/_Main/Scripts/UI/Settings/LanguageConfig.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 037da063161eeb141bc152101be5ee71
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
