--- Where_Is_My_Wife/Assets/_Main/Scripts/UI/Settings/UnityEventText.cs
@@ -1,19 +1,0 @@
-using System;
-using TMPro;
-using UnityEngine;
-using UnityEngine.UI;
-
-public class UnityEventText : MonoBehaviour
-{
-    [SerializeField] private TextMeshProUGUI _text;
-
-    public void SetText(string text)
-    {
-        _text.text = text;
-    }
-
-    public void SetText(float value)
-    {
-        _text.text = value.ToString();
-    }
-}
