+++ Where_Is_My_Wife/Assets/_Main/Scripts/UI/UnityEventText.cs
@@ -1,0 +1,20 @@
+using TMPro;
+using UnityEngine;
+
+namespace WhereIsMyWife.UI
+{
+    public class UnityEventText : MonoBehaviour
+    {
+        [SerializeField] private TextMeshProUGUI _text;
+
+        public void SetText(string text)
+        {
+            _text.text = text;
+        }
+
+        public void SetText(float value)
+        {
+            _text.text = value.ToString();
+        }
+    }
+}
